2_openai/deep_research - gemini/requirements.txt
# Core dependencies
gradio>=4.0.0
python-dotenv
pydantic>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
duckduckgo-search
sendgrid
requests
pypdf
openai-agents

# Note: The 'agents' library appears to be a custom library
# You may need to include it in your project or install from a specific source
